requirements.txt
requests
tqdm
pandas
openpyxl
python-dotenv
